changed datatype of audioStruct distance from integer to float as maximum and minimum is in range between 0...1

+++ Wiesn-Run/src/definitions.h
@@ -142,7 +142,7 @@ struct scoreStruct {
  */
 struct audioStruct {
     std::string name;
-    int distance;
+    float distance;
 };
 
 /**
